+++ src/backend/src/main/java/com/becon/opencelium/backend/database/mysql/service/ConnectionService.java
@@ -27,7 +27,6 @@ import com.github.fge.jsonpatch.JsonPatch;
 
 import java.io.IOException;
 import java.util.List;
-import java.util.Optional;
 
 public interface ConnectionService {
 
@@ -90,6 +89,4 @@ public interface ConnectionService {
     void deleteRule(long connectionId, long ruleId);
 
     void updateConnectionsToCurrentVersion();
-
-    void verifyMasterPassword(String masterPassword);
 }
