+++ src/backend/src/main/java/com/becon/opencelium/backend/database/mysql/service/ConnectorService.java
@@ -58,5 +58,7 @@ public interface ConnectorService {
 
     void update(Connector connector, Connector newConnector);
 
-    void updateRequestData(Integer id, Map<String, String> requestData);
+    void updateRequestData(Integer connectorId, Map<String, String> requestData);
+
+    void verifyMasterPassword(String masterPassword);
 }
